Added line to temp modify path
Added line that should allow most people to for go adding a new line to their path for GraphViz. You can still add the path yourself as any other program will still want it in your path. If the path variable in run.py is incorrect for you, change as needed.